Backend/requirements.txt.txt
asgiref-3.7.2
django-4.2.6 
sqlparse-0.4.4 
tzdata-2023.3
 psycopg2-binary-2.9.9
  psycopg2-2.9.9
 djangorestframework-3.14.0 
 pytz-2023.3.post1